Remove cookies

+++ server/DocumentsKM/Dtos/AdditionalWork/AdditionalWorkResponse.cs
@@ -1,11 +1,13 @@
+using System;
+
 namespace DocumentsKM.Dtos
 {
     public class AdditionalWorkResponse
     {
-        public int Id { get; set; }
+        public Int32 Id { get; set; }
         public EmployeeBaseResponse Employee { get; set; }
-        public int Valuation { get; set; }
-        public int MetalOrder { get; set; }
+        public Int16 Valuation { get; set; }
+        public Int16 MetalOrder { get; set; }
         public float DrawingsCompleted { get; set; }
         public float DrawingsCheck { get; set; }
     }
