+++ server/DocumentsKM/Dtos/MarkGeneralDataPoint/MarkGeneralDataPointResponse.cs
@@ -1,9 +1,11 @@
+using System;
+
 namespace DocumentsKM.Dtos
 {
     public class MarkGeneralDataPointResponse
     {
-        public int Id { get; set; }
+        public Int32 Id { get; set; }
         public string Text { get; set; }
-        public int OrderNum { get; set; }
+        public Int16 OrderNum { get; set; }
     }
 }
