+++ server/DocumentsKM/Dtos/MarkGeneralDataPoint/MarkGeneralDataPointUpdateRequest.cs
@@ -1,9 +1,11 @@
+using System;
+
 namespace DocumentsKM.Dtos
 {
     public class MarkGeneralDataPointUpdateRequest
     {
         public string Text { get; set; }
-        public int? OrderNum { get; set; }
+        public Int16? OrderNum { get; set; }
 
         public MarkGeneralDataPointUpdateRequest()
         {
