+++ server/DocumentsKM/Dtos/MarkLinkedDoc/MarkLinkedDocResponse.cs
@@ -1,10 +1,11 @@
+using System;
 using DocumentsKM.Models;
 
 namespace DocumentsKM.Dtos
 {
     public class MarkLinkedDocResponse
     {
-        public int Id { get; set; }
+        public Int16 Id { get; set; }
         public LinkedDoc LinkedDoc { get; set; }
         public string Note { get; set; }
     }
