+++ server/DocumentsKM/Dtos/Specification/SpecificationResponse.cs
@@ -4,8 +4,8 @@ namespace DocumentsKM.Dtos
 {
     public class SpecificationResponse
     {
-        public int Id { get; set; }
-        public int Num { get; set; }
+        public Int32 Id { get; set; }
+        public Int16 Num { get; set; }
         public bool IsCurrent { get; set; }
         public string Note { get; set; }
         public DateTime? CreatedDate { get; set; }
