+++ server/DocumentsKM/Models/DefaultValues.cs
@@ -1,3 +1,4 @@
+using System;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
@@ -10,26 +11,26 @@ namespace DocumentsKM.Models
         [ForeignKey("UserId")]
         public virtual User User { get; set; }
         [Key]
-        public int UserId { get; set; }
+        public Int16 UserId { get; set; }
 
         // Отдел
         [ForeignKey("DepartmentId")]
         public virtual Department Department { get; set; }
-        public int? DepartmentId { get; set; }
+        public Int16? DepartmentId { get; set; }
 
         // Разработчик
         [ForeignKey("CreatorId")]
         public virtual Employee Creator { get; set; }
-        public int? CreatorId { get; set; }
+        public Int32? CreatorId { get; set; }
 
         // Проверщик
         [ForeignKey("InspectorId")]
         public virtual Employee Inspector { get; set; }
-        public int? InspectorId { get; set; }
+        public Int32? InspectorId { get; set; }
 
         // Нормоконтролер
         [ForeignKey("NormContrId")]
         public virtual Employee NormContr { get; set; }
-        public int? NormContrId { get; set; }
+        public Int32? NormContrId { get; set; }
     }
 }
