+++ server/DocumentsKM/Models/Position.cs
@@ -1,3 +1,4 @@
+using System;
 using System.ComponentModel.DataAnnotations;
 
 namespace DocumentsKM.Models
@@ -6,7 +7,7 @@ namespace DocumentsKM.Models
     public class Position
     {
         [Key]
-        public int Id { get; set; }
+        public Int16 Id { get; set; }
 
         // Название
         [Required]
