+++ server/DocumentsKM/Models/User.cs
@@ -1,5 +1,6 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using System.ComponentModel.DataAnnotations;
+using System;
 
 namespace DocumentsKM.Models
 {
@@ -7,7 +8,7 @@ namespace DocumentsKM.Models
     public class User
     {
         [Key]
-        public int Id { get; set; }
+        public Int16 Id { get; set; }
 
         // Логин
         [Required]
@@ -22,6 +23,6 @@ namespace DocumentsKM.Models
         // Сотрудник
         [ForeignKey("EmployeeId")]
         public virtual Employee Employee { get; set; }
-        public int EmployeeId { get; set; }
+        public Int32 EmployeeId { get; set; }
     }
 }
